% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDevice.R
\name{calculated_feed_from_position}
\alias{calculated_feed_from_position}
\title{Calculate feed rate from the path position data items}
\usage{
calculated_feed_from_position(mtc_device, pattern = "PATH_POSITION")
}
\arguments{
\item{mtc_device}{is the MTCDevice object}

\item{pattern}{is the pattern of the path postion data items in the device object}
}
\description{
Returns a data.frame which contains the calculated feed rates and the corresponding time stamps
}
\examples{
data("example_mtc_device_3")
calculated_feed_from_postion = calculated_feed_from_position(example_mtc_device_3)
}

