% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{check_xml_log_mapping}
\alias{check_xml_log_mapping}
\title{Check if a device XML has been configured correctly}
\usage{
check_xml_log_mapping()
}
\arguments{
\item{data_from_log}{Device Adapter log in data frame form created using the 
\code{\link{read_adapter_log_file}} function}

\item{xpaths_map}{Device XML for this devicein data frame form created using the
\code{\link{get_xpaths_from_xml}} function}
}
\description{
If a device XML is configured correctly, there should be a one-to-one mapping
between the names of the data items from the log data and the names of 
data items configured in the devices XML. Only the ones that are mapped correctly
can be analysed further.
}
\details{
This function checks if the mapping is correct, and issues a warning if there are 
incorrect mappings
}
\examples{

  
}

