% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{clean_reduntant_vector}
\alias{clean_reduntant_vector}
\title{Remove Vector Redundancy}
\usage{
clean_reduntant_vector(data)
}
\arguments{
\item{data}{vector to be cleaned}
}
\description{
Replaces successive elements which are same, such that in result, no two adjacent values are the same.
}
\examples{
  clean_reduntant_vector(c(4,4,4,5,5))
}
