% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{create_mtc_device}
\alias{create_mtc_device}
\title{Create Device from different data sourcers}
\usage{
create_mtc_device(data_source = "adapter", ...)
}
\arguments{
\item{data_source}{Defines what the data source is}

\item{...}{Other arguments passed on to specific data creation functions}
}
\description{
This is a wrapper over the individual functions
}

