% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{create_mtc_device_from_adapter_data}
\alias{create_mtc_device_from_adapter_data}
\title{Create MTCDevice class from adapter data and log file}
\usage{
create_mtc_device_from_adapter_data(file_path_adapter_log, file_path_xml,
  device_name, mtconnect_version = NULL)
}
\arguments{
\item{file_path_adapter_log}{Path to adapter log file}

\item{file_path_xml}{Path to the Device XML file}

\item{device_name}{Name of the Device in the Device XML file}
}
\description{
Create MTCDevice class from adapter data and log file
}
\examples{
file_path_adapter_log = "tests/dataExtraction/test_log_data.log"
file_path_xml = "tests/dataExtraction/test_devices.xml"
device_name = "test_device"
mtc_device = create_mtc_device_from_adapter_data(
  system.file(file_path_adapter_log, package = "mtconnectR"),
  system.file(file_path_xml, package = "mtconnectR"),
  device_name)
print(summary(mtc_device))
}

