% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{create_mtcdevice_from_adapter_data}
\alias{create_mtcdevice_from_adapter_data}
\title{Create MTCDevice class from adapter data and log file}
\usage{
create_mtcdevice_from_adapter_data(file_path_adapter_log, file_path_xml,
  device_xml_name)
}
\arguments{
\item{file_path_adapter_log}{Path to adapter log file}

\item{file_path_xml}{Path to the Device XML file}

\item{device_xml_name}{Name of the Device in the Device XML file}
}
\description{
Create MTCDevice class from adapter data and log file
}

