% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{create_mtcdevice_from_agent_data}
\alias{create_mtcdevice_from_agent_data}
\title{title createDevicefromAgent}
\usage{
create_mtcdevice_from_agent_data(file_path_agent_log, device_uuid)
}
\description{
title createDevicefromAgent
}

