% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_gcode_parsed}
\alias{example_gcode_parsed}
\title{Example data set showing parsed G code data}
\format{A data frame with some rows and 8 variables}
\usage{
example_gcode_parsed
}
\description{
A manually created dataset showing a raw gcode data file, parsed and read into R. The columns are
\itemize{
  \item line  Line number
  \item single_block A single block of G code from a line
  \item value Value of the data item corresponding to the command
  \item priority Priority of the block as per the pre-written dictionary
  \item prefix Prefix of the block
  \item type Type 
  \item subtype Subtype
  \item supported Whether the specific G code block is supported or not by the dictionary
}
}
\keyword{datasets}

