% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_log_data}
\alias{example_log_data}
\title{Example data set showing MTC Log data}
\format{A data frame some rows and 3 variables}
\usage{
example_log_data
}
\description{
A manually created dataset showing a log data file, parsed and read into R. The columns are 
\itemize{
  \item timestamp. Timestamp of the event
  \item data_item_name Name of the data Item from the delimited MTC data. Can be empty.
  \item value of the data item
}
}
\keyword{datasets}

