% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_mapped_plot}
\alias{example_mapped_plot}
\title{ggplot object showing mapping between simulated and actual time series}
\format{An ggplot object}
\usage{
example_mapped_plot
}
\description{
ggplot object showing mapping between simulated and actual time series
}
\keyword{datasets}

