% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_mtc_device}
\alias{example_mtc_device}
\title{Example data set showing a MTConnect Device}
\format{An MTCDevice data item}
\usage{
example_mtc_device
}
\description{
The data can be accessed using the @ function. The slots are:
\itemize{
  \item rawdata Original delimited MTC data (parsed from the file using which the data was created)
  \item metadata Metadata (if any) for the device
  \item data_item_list Processed data showing each data item as a separate device
  \item device_uuid UUID of the device
}
}
\keyword{datasets}

