% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_mtc_device_2}
\alias{example_mtc_device_2}
\title{A bigger example data set showing a MTConnect Device with path position and conditions}
\format{An MTCDevice data item}
\usage{
example_mtc_device_2
}
\description{
The data can be accessed using the @ function. The slots are:
\itemize{
  \item rawdata Original delimited MTC data (parsed from which the data was created)
  \item metadata Metadata (if any) for the device
  \item data_item_list processed data showing each data item as a separate device
  \item device_uuid UUID of the device
}
}
\keyword{datasets}

