% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_mtc_device_sim}
\alias{example_mtc_device_sim}
\title{MTCDevice object showing simulated G code data}
\format{A data frame with some rows and 13 variables}
\usage{
example_mtc_device_sim
}
\description{
MTCDevice object created using the simulate_gcode function using parsed G code and 
convert_mtc_device_from_ts to convert data.frame to MTCDevice object.
}
\keyword{datasets}

