% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_mtc_sim_mapped}
\alias{example_mtc_sim_mapped}
\title{MTCDevice object containing actual and simulated data and the mapping}
\format{An MTCDevice object}
\usage{
example_mtc_sim_mapped
}
\description{
MTCDevice object containing actual and simulated data and the mapping
}
\keyword{datasets}

