% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_parsed_device_xml}
\alias{example_parsed_device_xml}
\title{Example dataset showing the parsed xml for a device}
\format{An MTCDevice data item}
\usage{
example_parsed_device_xml
}
\description{
The data can be accessed using the @ function. The slots are:
\itemize{
  \item parsed_xml Raw XML
  \item device_details Name,uuid and id of the device 
  \item mtconnect_version 
}
}
\keyword{datasets}
