% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_simulated_gcode_data}
\alias{example_simulated_gcode_data}
\title{Example data set showing simulated G code data}
\format{A data frame with some rows and 13 variables}
\usage{
example_simulated_gcode_data
}
\description{
Dataset created using the simulate_gcode function using parsed G code. The columns are
\itemize{
  \item timestamp Simulated timestamp
  \item lineid Line ID 
  \item program Program name
  \item tool_id Tool ID
  \item pfr Simulated path feed rate
  \item rot_vel Simulated rotational velocity
  \item x_pos Simulated X axis position
  \item y_pos Simulated Y axis position
  \item z_pos Simulated Z axis position
  \item x_vel Simulated X axis velocity
  \item y_vel Simulated Y axis velocity
  \item z_vel Simulated Z axis velocity
  \item state_upcoming_tool State upcoming tool
}
}
\keyword{datasets}

