% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_xpath_info}
\alias{example_xpath_info}
\title{Example data set showing Xpaths from a device XML}
\format{A data frame with some rows and 6 variables}
\usage{
example_xpath_info
}
\description{
Dataset showing a parsed DeviceXML file showing all the XPaths and the properties
\itemize{
  \item id ID of the data item
  \item name Name of the data Item from the delimited MTC data. Can be empty.
  \item type MTC Type of the data item
  \item category MTC Category of the data item
  \item subType MTC subType of the data item. Can be emoty
  \item xpath xpath showing the truncated path to the particular data item in the device XML
}
}
\keyword{datasets}

