% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtconnectR_staging.R
\name{filter_timestamps_mtc_device}
\alias{filter_timestamps_mtc_device}
\title{Filter MTCDevice object based on time range}
\usage{
filter_timestamps_mtc_device(mtc_device, start_time, end_time)
}
\arguments{
\item{mtc_device}{is the MTCDevice object}

\item{start_time}{is the Start time}

\item{end_time}{is the End time}
}
\description{
Helper function to quickly filter based on time range
}

