% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{find_line_type}
\alias{find_line_type}
\title{Returns the tine Type of s single line}
\usage{
find_line_type(lineRead)
}
\arguments{
\item{lineRead}{a single line from the adapter data}
}
\value{
The type of data either, ASSET or TS or COMMAND or UNKNOWN
}
\description{
Returns the tine Type of s single line
}

