% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{getData,MTCCycle-method}
\alias{getData,MTCCycle-method}
\title{Get Data from MTCDevice/MTCCycle Object as a data.frame}
\usage{
\S4method{getData}{MTCCycle}(.Object)
}
\arguments{
\item{.Object}{Object of MTCCycle or MTCDevice Class}
}
\description{
Get Data from MTCDevice/MTCCycle Object as a data.frame
}
\examples{
data("example_mtc_device")
getData(example_mtc_device)
}

