% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDataItem.R
\docType{methods}
\name{getData,MTCDataItem-method}
\alias{getData,MTCDataItem-method}
\title{Get data from the object in a data frame form}
\usage{
\S4method{getData}{MTCDataItem}(.Object)
}
\arguments{
\item{.Object}{A MTC Object}
}
\description{
Get data from the object in a data frame form
}
\examples{
data("example_mtc_data_item")
getData(example_mtc_data_item)
}
