% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDataItem.R
\name{getData}
\alias{getData}
\title{Get data from the object in a data frame form}
\usage{
getData(.Object, pattern)
}
\arguments{
\item{.Object}{An MTC Object}

\item{pattern}{OPTIONAL can be used to query specific data items}
}
\description{
Get data from the object in a data frame form
}
\examples{
data("example_mtc_data_item")
getData(example_mtc_data_item)
}

