% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{getDataItem,MTCCycle,ANY-method}
\alias{getDataItem,MTCCycle,ANY-method}
\title{Get the first dataitem}
\usage{
\S4method{getDataItem}{MTCCycle,ANY}(.Object)
}
\arguments{
\item{.Object}{Object of MTCCycle or MTCDevice Class}
}
\description{
Get the first dataitem
}
\examples{
data("example_mtc_device")
getDataItem(example_mtc_device)
}
