% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{getDataItem,MTCCycle,character-method}
\alias{getDataItem,MTCCycle,character-method}
\title{Get one or more data items from the MTCCycle or MTCDevice using a character pattern}
\usage{
\S4method{getDataItem}{MTCCycle,character}(.Object, pattern)
}
\arguments{
\item{.Object}{Object of MTCCycle or MTCDevice Class}

\item{pattern}{Regex of the pattern by which the data is queried}
}
\description{
Get one or more data items from the MTCCycle or MTCDevice using a character pattern
}
\examples{
data("example_mtc_device")
getDataItem(example_mtc_device, "POSIT")
}
