% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\name{getDataItem}
\alias{getDataItem}
\title{Get data on or more data items from the class}
\usage{
getDataItem(.Object, pattern)
}
\arguments{
\item{.Object}{object of MTCCycle or MTCDevice Class}

\item{pattern}{OPTIONAL can be used to query specific data items}
}
\description{
Get data on or more data items from the class
}
\examples{
data("example_mtc_device")
getDataItem(example_mtc_device)
}

