% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDataItem.R
\docType{methods}
\name{getMetaData,MTCDataItem-method}
\alias{getMetaData,MTCDataItem-method}
\title{Get MetaData from the Object as a list}
\usage{
\S4method{getMetaData}{MTCDataItem}(.Object)
}
\arguments{
\item{.Object}{Object of MTCDataItem Class}
}
\description{
Get MetaData from the Object as a list
}
\examples{
data("example_mtc_data_item")
getMetaData(example_mtc_data_item)
}
