% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDataItem.R
\name{getMetaData}
\alias{getMetaData}
\title{Get MetaData from the Object as a list}
\usage{
getMetaData(.Object)
}
\arguments{
\item{.Object}{Object of MTCDataItem Class}
}
\description{
Get MetaData from the Object as a list
}
\examples{
data("example_mtc_data_item")
getMetaData(example_mtc_data_item)
}
