% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gcode_mtc.R
\name{map_gcode_mtc}
\alias{map_gcode_mtc}
\title{Create a mapping between simulated and actual data}
\usage{
map_gcode_mtc(mtc_device_sim, mtc_device, elasticity = 2)
}
\arguments{
\item{mtc_device_sim}{is the simlated version}

\item{mtc_device}{is the actual log data}

\item{elasticity}{is the maximum consecutive reference elements skippable (passed to dtw::mvmStepPattern())}
}
\description{
Creates a timestamp based mapping to map every simulated timestamp to an actual timestamp based on real data
}

