% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{merge,MTCCycle,ANY-method}
\alias{merge,MTCCycle,ANY-method}
\title{Merge all data items from the MTCCycle or MTCDevice}
\usage{
\S4method{merge}{MTCCycle,ANY}(x)
}
\arguments{
\item{x}{Object of MTCCycle or MTCDevice Class}
}
\description{
Merge all data items from the MTCCycle or MTCDevice
}
\examples{
data("example_mtc_device")
merge(example_mtc_device)
}
