% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{merge,MTCCycle,numeric-method}
\alias{merge,MTCCycle,numeric-method}
\title{Merge on or more data items from the MTCCycle or MTCDevice using an index}
\usage{
\S4method{merge}{MTCCycle,numeric}(x, y)
}
\arguments{
\item{.Object}{object of MTCCycle or MTCDevice Class}
}
\description{
Merge on or more data items from the MTCCycle or MTCDevice using an index
}
\examples{
data("example_mtc_device")
merge(example_mtc_device, 1)
}

