% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{class}
\name{MTCCycle-class}
\alias{MTCCycle-class}
\title{An S4 class to represent the different data items of a device}
\description{
An S4 class to represent the different data items of a device
}
\section{Slots}{

\describe{
\item{\code{data_item_list}}{List of data items with data(data.frame of timestamp,value),
data_type,path,dataSource,xmlID}

\item{\code{device_uuid}}{UUID of the device}
}}

