% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDevice.R
\docType{class}
\name{MTCDevice-class}
\alias{MTCDevice-class}
\title{An S4 class to represent a device.}
\description{
An S4 class to represent a device.
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{Delimited MTC data (parsed from the file using which the data was created)}

\item{\code{metadata}}{Metadata (if any about) the device}

\item{\code{data_item_list}}{List of data items with data(data.frame of timestamp,value),data_type,path,dataSource,xmlID}

\item{\code{device_uuid}}{UUID of the device}
}}
\examples{
data('example_mtc_data_item')
mtc_data_item <- example_mtc_data_item
device_name = extract_param_from_xpath(mtc_data_item@path,"Device")
mtc_device_object = list(rawdata = list(),metadata = list(),data_item_list = list(mtc_data_item),device_uuid = device_name)

}

