% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{read_adapter_log_file}
\alias{read_adapter_log_file}
\title{Function to load the log data into R as a data.frame}
\usage{
read_adapter_log_file(file_path_adapter_log, condition_names = c())
}
\arguments{
\item{file_path_adapter_log}{Path to the file containing log data}

\item{condition_names}{A character string with the names of the data items that
represents the conditions in the log data}
}
\description{
Function to load the log data into R as a data.frame
}
\examples{
device_name = "test_device"
file_path_xml = "testdata/dataExtraction/test_devices.xml"
xpath_info = get_xpaths_from_xml(system.file(file_path_xml, package = "mtconnectR"), device_name)
}

