% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction_mtc.R
\name{read_dmtcd_file}
\alias{read_dmtcd_file}
\title{Function to load Log data into R as a data.frame}
\usage{
read_dmtcd_file(file_path_dmtcd, condition_names = c(),
  path_position_names = c())
}
\arguments{
\item{file_path_dmtcd}{Path to the file containing log data}

\item{condition_names}{A character string with the names of the data items that
represent the conditions in the log data}

\item{path_position_names}{A character string with the names of the data items that
represent the path_position data items}
}
\description{
Function to load Log data into R as a data.frame
}
\examples{
device_name = "test_device"
file_path_xml = "testdata/dataExtraction/test_devices.xml"
xpath_info = get_xpaths_from_xml(system.file(file_path_xml, package = "mtconnectR"), device_name)
}
