% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{sequence_order_vector}
\alias{sequence_order_vector}
\title{Sequence Order Vector.}
\usage{
sequence_order_vector(data)
}
\arguments{
\item{data}{The data to generate ids for}
}
\description{
Create IDs for a vector. Successive elements if same, are given the same ID.
}
\examples{
 sequence_order_vector(c(4,4,4,5,5))
}
\seealso{
\code{\link{rle}}, \code{\link{clean_reduntant_vector}}
}
