% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gcode.R
\name{simulate_data_from_gcode}
\alias{simulate_data_from_gcode}
\title{Simulate position,velocity and other data from G-code}
\usage{
simulate_data_from_gcode(gcode_parsed, start_time = 0, data_res = 0.2,
  data_type = "ISO")
}
\arguments{
\item{gcode_parsed}{Parsed gcode}

\item{start_time}{Starting time (default 0)}

\item{data_res}{Resolution for occurrence of a new data point. (seconds)}

\item{data_type}{Data type}
}
\description{
Reads parsed gcode and returns simulated data
}
\examples{
data("example_gcode_parsed")
simulated_data_from_gcode <- simulate_data_from_gcode(example_gcode_parsed,start_time = 0, 
data_res = 0.2, data_type = "HH")
}
