% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{summary,MTCCycle-method}
\alias{summary,MTCCycle-method}
\title{Show a quick summary of the MTCCyle or MTCDevice Object}
\usage{
\S4method{summary}{MTCCycle}(object)
}
\arguments{
\item{object}{The MTCDevice or MTCCycle object}
}
\description{
Show a quick summary of the MTCCyle or MTCDevice Object
}

