% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDataItem.R
\docType{methods}
\name{summary,MTCDataItem-method}
\alias{summary,MTCDataItem-method}
\title{Show a quick summary of the MTCDataItem}
\usage{
\S4method{summary}{MTCDataItem}(object)
}
\arguments{
\item{object}{The MTCDataItem object}
}
\description{
Show a quick summary of the MTCDataItem
}
\examples{
data("example_mtc_data_item")
summary(example_mtc_data_item)
}

