% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDevice.R
\name{ts_to_mtc_device}
\alias{ts_to_mtc_device}
\title{Create an MTC device object from a merged time series data frame}
\usage{
ts_to_mtc_device(merged_device, device_uuid = "unmerged_device")
}
\arguments{
\item{merged_device}{An existing object of MTCDevice Class}

\item{device_uuid}{UUID to be given to the device}
}
\description{
Create an MTC device object from a merged time series data frame
}
\examples{
data("example_mtc_device")
merged_device = merge(example_mtc_device)
ts_to_mtc_device(merged_device)

}

