% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{create_device}
\alias{create_device}
\title{Create Device from different data sourcers}
\usage{
create_device(data_source = "adapter", ...)
}
\description{
Create Device from different data sourcers
}

