% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{create_mtc_device_from_adapter_data}
\alias{create_mtc_device_from_adapter_data}
\title{Create MTCDevice class from adapter data and log file}
\usage{
create_mtc_device_from_adapter_data(file_path_adapter_log, file_path_xml,
  device_xml_name)
}
\arguments{
\item{file_path_adapter_log}{Path to adapter log file}

\item{file_path_xml}{Path to the Device XML file}

\item{device_xml_name}{Name of the Device in the Device XML file}
}
\description{
Create MTCDevice class from adapter data and log file
}
\examples{
file_path_adapter_log = "extdata/tft-405-pfh.log"
file_path_xml = "extdata/Devices.xml.txt"
device_xml_name = "TFT-405-PFH"
mtc_device = create_mtc_device_from_adapter_data(
  system.file(file_path_adapter_log, package = "mtconnectR"),
  system.file(file_path_xml, package = "mtconnectR"),
  device_xml_name)
print(summary(mtc_device))
}

