% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{example_mtc_data_item}
\alias{example_mtc_data_item}
\title{Example data set showing a MTConnect DataItem}
\format{An MTCDevice data item}
\usage{
example_mtc_data_item
}
\description{
The data can be accessed using the @ function. The slots are:
\itemize{
  \item data Data for a single data item in a data.frame in timestamp, value format
  \item data_type Type of Data - can be event or sample
  \item path XML Xpath
  \item data_source Source from which the data item was created
  \item xmlID ID of the data item in the devices XML
}
}
\keyword{datasets}

