% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDevice.R
\name{filter_timestamps_mtc_device}
\alias{filter_timestamps_mtc_device}
\title{Filter MTCDevice object based on time range}
\usage{
filter_timestamps_mtc_device(mtc_device, start_time, end_time)
}
\arguments{
\item{mtc_device}{is the MTCDevice objectcc}

\item{start_time}{is the Start time}

\item{end_time}{is the End time}
}
\description{
Helper function to quickly filter based on time range
}
\examples{
data("example_mtc_device_3")
start_time = as.POSIXct("2016-03-22 12:45:00.000")
end_time = as.POSIXct("2016-03-22 12:45:10.000")
filtered_data = filter_timestamps_mtc_device(example_mtc_device_3,start_time,end_time)
}
