% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDataItem.R
\docType{methods}
\name{getData,MTCDataItem-method}
\alias{getData,MTCDataItem-method}
\title{Get Data from the Object as a data.frame}
\usage{
\S4method{getData}{MTCDataItem}(.Object)
}
\arguments{
\item{.Object}{Object of MTCDataItem Class}
}
\description{
Get Data from the Object as a data.frame
}
\examples{
data("example_mtc_data_item")
getData(example_mtc_data_item)
}

