% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{merge,MTCCycle,character-method}
\alias{merge,MTCCycle,character-method}
\title{Merge on or more data items from the MTCCycle or MTCDevice using a character pattern}
\usage{
\S4method{merge}{MTCCycle,character}(x, y)
}
\arguments{
\item{x}{object of MTCCycle or MTCDevice Class}

\item{y}{regexp of the pattern by which the merge has to be done}
}
\description{
Merge on or more data items from the MTCCycle or MTCDevice using a character pattern
}
\examples{
data("example_mtc_device")
merge(example_mtc_device, "POSIT")
}

