% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCCycle.R
\docType{methods}
\name{merge,MTCCycle,character-method}
\alias{merge,MTCCycle,character-method}
\title{Merge one or more data items from the MTCCycle or MTCDevice using a character pattern}
\usage{
\S4method{merge}{MTCCycle,character}(x, y)
}
\arguments{
\item{x}{Object of MTCCycle or MTCDevice Class}

\item{y}{Regex for picking data items which has to be merged}
}
\description{
Merge one or more data items from the MTCCycle or MTCDevice using a character pattern
}
\examples{
data("example_mtc_device")
merge(example_mtc_device, "POSIT")
}
