% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeTS}
\alias{mergeTS}
\title{Merges all the data.frames in the list into single data.frame}
\usage{
mergeTS(DF_list, output_DF = T, use_list_names = F,
  additional_ts = .POSIXct(integer(0)), ignore_tz = F)
}
\arguments{
\item{DF_list}{is a list of data.frames. Each data.frame should be of type timestamp|value1|value2...}

\item{output_DF}{if TRUE, then returns output in the form of data.frame instead of data.table}

\item{use_list_names}{if TRUE, the names of the list are assigned the columns names}

\item{additional_ts}{an POSIXct vector of timestamps which needs to be added into the table.
The values are repeated from the previous timestamp}

\item{ignore_tz}{timezone error is ignored.}
}
\description{
Merges all the data.frames in the list into single data.frame
}
\examples{
data("example_mtc_device")
mtc_device = merge(example_mtc_device)
df_1 = mtc_device[seq(1,17,2),]
df_2 = mtc_device[seq(2,17,2),]
merged_df = mergeTS(list(df_1,df_2))
}
