% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCDevice.R
\docType{class}
\name{MTCDevice-class}
\alias{MTCDevice-class}
\title{An S4 class to represent a device. It contains MTCCycle class and the slots below.}
\description{
An S4 class to represent a device. It contains MTCCycle class and the slots below.
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{Delimited MTC data (parsed from the file using which the data was created)}

\item{\code{metadata}}{Metadata (if any about) the device}
}}

\examples{
data('example_mtc_data_item')
example_mtc_data_item
}
